api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfminer.six 
google-cloud-aiplatform  # for Gemini
openai